Fix merge errors and circular import
